+++ tomcat/src/test/java/com/techcourse/controller/LoginControllerTest.java
@@ -1,4 +1,4 @@
-package com.techcourse.handler;
+package com.techcourse.controller;
 
 import jakarta.http.Header;
 import jakarta.http.HttpHeaderKey;
